[Java Package] FQCN 과 Built-in packages
