+++ src/main/java/com/oh29oh29/study03/sample01/Problem.java
@@ -1,11 +1,5 @@
 package com.oh29oh29.study03.sample01;
 
-import static com.oh29oh29.study03.sample02.Java.KEYWORD;
-import static com.oh29oh29.study03.sample02.Java.print;
-
 public class Problem {
-    public static void main(String[] args) {
-        System.out.println("Java keyword[0]: " + KEYWORD[0]);
-        print();
-    }
+    com.oh29oh29.study02.sample02.Java java = new com.oh29oh29.study02.sample02.Java();
 }
