+++ src/main/java/com/oh29oh29/study03/sample02/Java.java
@@ -1,10 +1,4 @@
 package com.oh29oh29.study03.sample02;
 
 public class Java {
-
-    public static String[] KEYWORD = {"package", "import"};
-
-    public static void print() {
-        System.out.println("Static Import...");
-    }
 }
