+++ src/main/java/com/oh29oh29/study04/sample02/Java.java
@@ -1,0 +1,10 @@
+package com.oh29oh29.study04.sample02;
+
+public class Java {
+
+    public static String[] KEYWORD = {"package", "import"};
+
+    public static void print() {
+        System.out.println("Static Import...");
+    }
+}
